reservas/serializers/serializers.py
from rest_framework import serializers
from ..models import Hotel, Reservation

class HotelSerializer(serializers.ModelSerializer):
    class Meta:
        model = Hotel
        fields = '__all__'

class ReservaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Reservation
        fields = '__all__'